feat: Update Parking Contract and Cleanup README
Updates all Smart Contracts to use Solidity version 0.8.23 from version 0.8.0 and makes minor changes in the behavior of parking sessions. Updates the `package.json` and `package-lock.json` files for the `truffle` package. These changes aim to make the code cleaner and more maintainable in addition to ensuring it can be deployed to Ethereum public testnets such as Sepolia