feat: Add Alchemy Sepolia Node Support and Downgrade Solidity Version
This commit introduces support for the Alchemy Sepolia node in the Truffle project, enhancing our ability to interact with the Sepolia testnet. Additionally, the Solidity version has been downgraded from 0.8.23 to 0.8.19. This change addresses a compatibility issue with Ganache, where a new EVM OpCode introduced in later Solidity versions was causing deployment failures. By reverting to Solidity 0.8.19, we maintain compatibility with Ganache's current deployment capabilities.

For more details on the Ganache issue with newer Solidity versions, refer to:
https://ethereum.stackexchange.com/questions/155718/invalid-opcode-while-deploying-helloworld